Concatenate together adjacent runs of PlainText together when parsing [Inline]
Since we want to parse in-line URLs (HTTP(S) and angle-delimited), and we want those URLs to be parsed only as within word boundaries, we need to parse words and whitespace separately. But I don't want to have `Block`s where whitespace and words are separate `Inline`s, as that'll take up unnecessary space, so we want to concatenate them together. This means that within our `Block`s that store `[Inline]` (`Paragraph`, `Bullet`, `Quote`), no two consecutive `Inline` elements will both be `PlainText`.