Rename `Block` to `AuthoredBlock`, rename `Block` data constructors with A prefix (for Authored), rename Document to AuthoredDocument
The `AuthoredBlock` and `AuthoredDocument` types refer to how a Subtext document is written by some Author. However, now that we are going to be supporting transclusion, how a Subtext document is written is not necessarily the same as how it's going to be read, since we'll need to fill the transclusion sections with the actual transclusion content.