Remove `Document` type synonym for `[Block]`
I think the synonym is just getting in the way of my own mental understanding of the Subtextual behaviour I want. I think I want to create `Document` into a type constructor that combines 1. a DocumentName and 2. the list of blocks. That'll let me do things like process a list of files into a `[Document Authored]` and then process that list into a graph so that I can resolve transclusions, creating a `[Document Readable]`.